Refactor : 장소 좋아요 부분 파일 분리
장소 좋아요 부분을 PlaceLike 파일로 분리하고
장소 text 부분도 PlaceText 파일로 분리하였다.
