+++ client/project-there/src/components/journeyDetail/PlaceDetailInfo/index.jsx
@@ -1,11 +1,8 @@
-import React, { useRef, useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import styled from 'styled-components';
-import { AiOutlineHeart, AiFillHeart } from 'react-icons/ai';
-import { authAxios } from '../../../lib/utills/customAxios';
-import { useAuthValue } from '../../../contexts/auth';
 import PlaceCommentList from './PlaceCommentList';
-
-const contentHeight = (condition, then, otherwise) => (condition ? then : otherwise);
+import PlaceLike from './PlaceLike';
+import PlaceText from './PlaceText';
 
 const StyledInfoHolder = styled.div`
   width: 50%;
@@ -20,184 +17,25 @@ const StyledTitle = styled.h2`
   letter-spacing: -4%;
   margin-bottom: 12%;
 `;
-const StyeldContents = styled.p`
-  font-weight: 300 !important;
-  font-size: 14px !important;
-  letter-spacing: -4% !important;
-  height: ${(props) => (props.displayOverFlowed ? '58%' : contentHeight(props.overflowed, '20%', '25%'))};
-  white-space: pre-wrap;
-  text-overflow: ${(props) => !props.displayOverFlowed && 'ellipsis'};
-  overflow: ${(props) => (props.displayOverFlowed ? 'auto' : 'hidden')};
-  word-break: break-all;
-  display: ${(props) => !props.displayOverFlowed && '-webkit-box'};
-  -webkit-line-clamp: 4;
-  -webkit-box-orient: vertical;
-  border-bottom: ${(props) => !props.overflowed && '1px solid #bcc4c6'};
-`;
-const StyledContentDetail = styled.div`
-  color: #bcc4c6;
-  font-size: 12px;
-  font-weight: 500;
-  width: 100%;
-  position: ${(props) => (props.displayCommentOverflowed ? 'absolute' : 'relative')};
-  top:${(props) => (!props.displayCommentOverflowed && '10px')};
-  bottom:${(props) => (props.displayCommentOverflowed && '14%')};
-  cursor: pointer;
-  background-color: var(--color-gray100);
-  height:20px;
-  border-bottom: 1px solid var(--color-gray400);
-`;
-
-const StyledLikeHolder = styled.div`
-  padding: 0 0 5px 0;
-  display: flex;
-  align-items: center;
-  span{
-    margin-left: 6px;
-    font-size: var(--font-regular);
-    height:25px;
-    font-weight: var(--weight-bold);
-    padding-top: 2px;
-  }
-  margin-bottom: 10px;
-`;
-const StyledLikeIconHolder = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  cursor: pointer;
-  svg{
-    width: 20px;
-    height: 20px;
-    color: red;
-  }
-`;
 
 export default function PlaceDetailInfo({ target, setTotalPlacesData }) {
-  const [isOverflowed, setIsOverFlowed] = useState(false);
-  const [displayOverflowed, setDisplayOverFlowed] = useState(false);
-  const [likeCount, setLikeCount] = useState(0);
-  const [isLiked, setIsLiked] = useState(false);
-  const memberInfo = useAuthValue();
-  const textRef = useRef();
-  const likeRef = useRef(isLiked);
-
-  // 좋아요 표시한 부분에 변경이 있었는지 확인해주는 함수
-  const checkLiked = () => {
-    if (likeRef.current !== target.heartedCheck) {
-      if (likeRef.current) {
-        // 서버에 해당 사실 날리고
-        const url = `place/heart?placeId=${target.placeId}`;
-        authAxios.post(url).catch((err) => { console.log(err); });
-        // totalPlacesData 수정해줌
-        setTotalPlacesData((prev) => prev.map((place) => {
-          if (place.placeId === target.placeId) {
-            const newPlace = { ...place };
-            newPlace.heartedCheck = true;
-            const { placeHeartCount } = newPlace;
-            if (placeHeartCount.indexOf('k') === -1) {
-              newPlace.placeHeartCount = String(Number(placeHeartCount) + 1);
-            }
-            return newPlace;
-          }
-          return place;
-        }));
-      } else {
-        const url = `place/unheart?placeId=${target.placeId}`;
-        authAxios.post(url).catch((err) => { console.log(err); });
-        // totalPlacesData 수정해줌
-        setTotalPlacesData((prev) => prev.map((place) => {
-          if (place.placeId === target.placeId) {
-            const newPlace = { ...place };
-            newPlace.heartedCheck = false;
-            const { placeHeartCount } = newPlace;
-            if (placeHeartCount.indexOf('k') === -1) {
-              newPlace.placeHeartCount = String(Number(placeHeartCount) - 1);
-            }
-            return newPlace;
-          }
-          return place;
-        }));
-      }
-    }
-  };
-
-  useEffect(() => {
-    if (textRef && !displayOverflowed) {
-      const overFlowCheck = textRef.current.scrollHeight > textRef.current.clientHeight;
-      setIsOverFlowed(overFlowCheck);
-    }
-  }, [textRef, displayOverflowed]);
-
-  // 좋아요 데이터 받아오기
-  useEffect(() => {
-    // 좋아요 수를 반영해줌
-    console.log(target);
-    const { placeHeartCount } = target;
-    if (placeHeartCount.indexOf('k') === -1) {
-      setLikeCount(Number(placeHeartCount));
-    } else {
-      setLikeCount(placeHeartCount);
-    }
-    // 좋아요 유무를 반영해줌
-    if (target.heartedCheck) {
-      setIsLiked(true);
-    }
-    return checkLiked;
-  }, []);
-
-  // 마지막에 좋아요 바뀌었는지 유무를 보내주기 위해 ref 업데이트를 해주는 부분
-  useEffect(() => {
-    likeRef.current = isLiked;
-  }, [isLiked]);
-
-  const handleContentClick = () => {
-    setDisplayOverFlowed((pre) => !pre);
-  };
-  const handleLikeButtonClick = () => {
-    if (!memberInfo?.nickName) {
-      window.alert('로그인이 필요한 서비스입니다.');
-    } else {
-      setIsLiked((pre) => !pre);
-      if (typeof (likeCount) === 'number') {
-        if (isLiked) {
-          setLikeCount((pre) => pre - 1);
-        } else {
-          setLikeCount((pre) => pre + 1);
-        }
-      }
-    }
-  };
-
+  const [isTextOverflowed, setIsTextOverFlowed] = useState(false);
+  const [displayTextOverflowed, setDisplayTextOverFlowed] = useState(false);
   return (
     <StyledInfoHolder>
-      <StyledLikeHolder>
-        <StyledLikeIconHolder
-          onClick={handleLikeButtonClick}
-        >
-          {isLiked ? <AiFillHeart /> : <AiOutlineHeart />}
-        </StyledLikeIconHolder>
-        <span>
-          {likeCount}
-        </span>
-      </StyledLikeHolder>
+      <PlaceLike
+        setTotalPlacesData={setTotalPlacesData}
+        target={target}
+      />
       <StyledTitle>{target.title}</StyledTitle>
-      <StyeldContents
-        ref={textRef}
-        displayOverFlowed={displayOverflowed}
-        overflowed={isOverflowed}
-      >
-        {target.text}
-      </StyeldContents>
-      {!!isOverflowed && (
-        <StyledContentDetail
-          onClick={handleContentClick}
-        >
-          {!displayOverflowed ? '더보기' : '접기'}
-        </StyledContentDetail>
-      )}
+      <PlaceText
+        text={target.text}
+        controlOverflow={[isTextOverflowed, setIsTextOverFlowed]}
+        controlDisplay={[displayTextOverflowed, setDisplayTextOverFlowed]}
+      />
       <PlaceCommentList
-        isOverflowed={isOverflowed}
+        isTextOverflowed={isTextOverflowed}
+        isTextDisplayOverflowed={displayTextOverflowed}
         placeId={target?.placeId}
       />
     </StyledInfoHolder>
